add: #2280 remove some dynamic imports and fix spinner logic
